pyproject.toml
[project]
name = "hackeurope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "abnf==2.2.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "asyncio==4.0.0",
    "attrs==25.4.0",
    "backoff==2.2.1",
    "cachetools==7.0.1",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "chardet==5.2.0",
    "charset-normalizer==3.4.4",
    "cint==1.0.0",
    "click==8.3.1",
    "cryptography==46.0.5",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "dotenv-python==0.0.1",
    "fickling==0.1.8",
    "filelock==3.24.3",
    "frozenlist==1.8.0",
    "gitdb==4.0.12",
    "gitpython==3.1.46",
    "google-auth==2.48.0",
    "google-genai==1.64.0",
    "gql==4.0.0",
    "graphql-core==3.2.7",
    "graphviz==0.21",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "intervaltree==3.2.1",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "kaitaistruct==0.11",
    "logging==0.4.9.6",
    "markupsafe==3.0.3",
    "multidict==6.7.1",
    "networkx==3.6.1",
    "numpy==2.4.2",
    "ollama==0.6.1",
    "packaging==26.0",
    "pandas==3.0.1",
    "pdfminer-six==20260107",
    "pillow==12.1.1",
    "platformdirs==4.9.2",
    "polyfile-weave==0.5.9",
    "propcache==0.4.1",
    "protobuf==6.33.5",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "requests==2.32.5",
    "rpds-py==0.30.0",
    "rsa==4.9.1",
    "scikit-learn==1.8.0",
    "scipy==1.17.0",
    "sentry-sdk==2.53.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "tenacity==9.1.4",
    "threadpoolctl==3.6.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "wandb==0.25.0",
    "weave==0.52.28",
    "websockets==15.0.1",
    "yarl==1.22.0",
]
